// src/app/master/constants/enums/yarn-type.enum.ts
export enum YarnTypeCountEnum {
    COUNT = "C",
    DENIER = "D",
    LINEN = "L"
}

export enum YarnTypeEnum {
    COTTON = "C",
    POYESTER = "P",
    TEXTURE = "T",
    WOOL = "W",
    RAYONVISOSE = "R",
    VISCOSERAYON = "V",
    LYCRA = "Y",
    LINEN = "L",
    TEXTYARN = "X",
    JUTE = "J",
    FANCY = "F",
    BLENDYARN = "B",
    OTHERYARN = "O",
    NYLON = "N",
}